requirements.txt
# Web 服务
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
python-dotenv>=1.0.0
jinja2>=3.1.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# AI API
openai>=1.12.0

# 企业微信加解密
pycryptodome>=3.20.0

# HTTP 客户端（企业微信主动发消息）
httpx>=0.27.0

# 原有依赖
Pillow>=9.0.0
